Add files via upload
This is an example I made at home which detects remotes instead of parking lines to create a space between using a calibration phase at the start of the program. At that point using a cup you can fill in the space and it will be detected as an occupied space.